+++ src/nespresso/bot/creator.py
@@ -5,3 +5,6 @@ from nespresso.core.configs.env_reader import settings
 bot = Bot(token=settings.TELEGRAM_BOT_TOKEN.get_secret_value())
 
 dp = Dispatcher()
+
+
+BOT_ID = bot.id