+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.1.0] - 2025-05-22 - Vlad
 ### Added
 - Send notification to user in case of error
+- Clear user's state when error occurs
 
 ## [0.1.0] - 2025-05-21 - Vlad
 ### Updated